+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonIntegratedToolbarQuickPrintCommand.cs
@@ -70,7 +70,7 @@ namespace Krypton.Toolkit
         {
             _imageStates = new ButtonImageStates();
 
-            Text = KryptonLanguageManager.ToolBarStrings.QuickPrint;
+            Text = KryptonManager.Strings.IntegratedToolBarStrings.QuickPrint;
         }
 
         #endregion